separated classes in task 2
